refactor(routes): extract /spend into spend.js (god-file decomposition)
Route contract unchanged (284 routes), 0 /spend left inline, vitest 38 passed 268 passed . Lucy (GLM-5.2), driver-gated.
